Create HTML-CSS(YouTube)
Hello everyone,

I'm thrilled to share that I have successfully completed my inaugural project, which involved the creation of a YouTube-style platform utilizing HTML and CSS. 
Despite initial challenges, the entire process proved to be an enjoyable and enlightening learning journey.

The crux of the project revolved around the implementation of various interactive elements, such as the search bar and voice search features within the platform's header. 
Upon hovering over these elements, users are able to input their responses. Furthermore, a sidebar containing essential links such as "home" was integrated into the design. As an additional feature, when the window size is reduced, two videos are displayed alongside the persistent header, ensuring a seamless and responsive user experience.

As I bring two weeks of intense learning to a close with the successful completion of this engaging and complex project, 
I eagerly anticipate the prospect of embracing new challenges in upcoming projects.